+++ changelog.md
@@ -7,6 +7,7 @@ Features
 * More complete and up-to-date set of MySQL reserved words for completions.
 * Place exact-leading completions first.
 * Allow history file location to be configured.
+* Make destructive-warning keywords configurable.
 
 
 Bug Fixes